src/client/app/vk-post-commentators.component.ts
import {Component} from 'angular2/core';
import {RouteConfig, ROUTER_DIRECTIVES, ROUTER_PROVIDERS} from 'angular2/router';

@Component({
  moduleId: __moduleName,
  selector: 'vk-post-commentators-app',
  providers: [ROUTER_PROVIDERS],
  templateUrl: 'vk-post-commentators.component.html',
  styleUrls: ['vk-post-commentators.component.css'],
  directives: [ROUTER_DIRECTIVES],
  pipes: []
})
@RouteConfig([
])
export class VkPostCommentatorsApp {
  defaultMeaning: number = 42;

  meaningOfLife(meaning?: number) {
    return `The meaning of life is ${meaning || this.defaultMeaning}`;
  }
}
